Match only perl6 source files
... when looking for files to trigger a new test run.  The tests note
that only .pm and .t files should be tracked.  The code was actually
matching all files under `lib/` and `t/`, which meant that precomp files
and editor backup files (e.g. vim ~ files) were all being matched.
This change ensures that only perl6 source files are matched.
Currently, files with the extensions .pm, .pm6 and .t are common, hence
this is what is implemented here.  The extensions .pl, .p6 and .pl6
have been omitted since they are usually associated with stand-alone
scripts, whereas this project seems to want to run the tests if the
library has changed, hence the rationale for this omission.